one hot labels of tensors to multi label classiication

+++ analysis/BERT_training_sliding_window.py
@@ -66,7 +66,7 @@ def convert_labels(labels):
     label_map = {"1": 0, "2": 1, "3": 2}
     labels_conv = [label_map[label] for label in labels]
     print("labels converted")
-    labels_onehot = torch.nn.functional.one_hot(labels_conv, num_classes=3)
+    labels_onehot = torch.nn.functional.one_hot(labels_conv, num_classes=3).float()
     return labels_onehot
